feat: Adiciona entidade e migração de categorias

+++ BackEnd/src/Gafel.Infrastructure/Migrations/Versions/Version0000003.cs
@@ -10,7 +10,7 @@ public class Version0000003 : VersionBase
         CreateTableWithDefaults(DatabaseTables.DEFAULT_CATEGORIES)
             .WithColumn("name").AsString(100).NotNullable()
             .WithColumn("is_active").AsBoolean().NotNullable()
-            .WithColumn("type").AsInt64().NotNullable()
+            .WithColumn("type").AsInt32().NotNullable()
             .WithColumn("icon").AsString(50).NotNullable();
     }
 }
